+++ integration/houdini.config.js
@@ -7,6 +7,7 @@ const config = {
   apiUrl: 'http://localhost:4000/graphql',
   defaultCachePolicy: 'CacheOrNetwork',
   defaultPartial: true,
+  routes: (filepath) => filepath && !filepath.includes('_') && !filepath.includes('+'),
   scalars: {
     DateTime: {
       type: 'Date',