+++ src/runtime/inline/query.ts
@@ -1,15 +1,7 @@
 // externals
-import { derived, get, Readable, writable } from 'svelte/store'
+import { derived, Readable } from 'svelte/store'
 // locals
-import { marshalInputs } from '../lib/scalars'
-import {
-	GraphQLTagResult,
-	Operation,
-	QueryResult,
-	CachePolicy,
-	TaggedGraphqlQuery,
-} from '../lib/types'
-import { getPage, getSession, goTo } from '../adapter'
+import { GraphQLTagResult, Operation, QueryResult, CachePolicy } from '../lib/types'
 import { wrapPaginationStore, PaginatedDocumentHandlers } from '../lib/pagination'
 import { getHoudiniContext } from '../lib/context'